Chore: rename app; add wsgi
